separate agent and server binary
