+++ .drone.sh
@@ -22,4 +22,5 @@ set -x
 git clone git@github.com:drone/drone-enterprise.git extras
 
 # build a static binary with the build number and extra features.
-go build -ldflags '-extldflags "-static" -X github.com/drone/drone/version.VersionDev=build.'${DRONE_BUILD_NUMBER} -tags extras -o release/drone github.com/drone/drone/drone
+go build -ldflags '-extldflags "-static" -X github.com/drone/drone/version.VersionDev=build.'${DRONE_BUILD_NUMBER} -o release/drone-server github.com/drone/drone/extras/cmd/drone-server
+go build -ldflags '-extldflags "-static" -X github.com/drone/drone/version.VersionDev=build.'${DRONE_BUILD_NUMBER} -o release/drone-agent  github.com/drone/drone/cmd/drone-agent
