+++ Dockerfile.agent
@@ -1,0 +1,7 @@
+# docker build --rm -t drone/drone .
+
+FROM centurylink/ca-certs
+
+ADD release/drone-agent /bin/
+
+ENTRYPOINT ["/bin/drone-agent"]
