--- Dockerfile.windows
@@ -1,11 +1,0 @@
-FROM ??????????
-
-EXPOSE 8000
-ENV DATABASE_DRIVER=sqlite3
-ENV DATABASE_CONFIG=/var/lib/drone/drone.sqlite
-ENV GODEBUG=netdns=go
-
-ADD release/windows/amd64/drone /drone
-
-ENTRYPOINT ["/drone"]
-CMD ["server"]
