+++ cmd/drone-server/main.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/drone/drone/version"
+
+	_ "github.com/joho/godotenv/autoload"
+	"github.com/urfave/cli"
+)
+
+func main() {
+	app := cli.NewApp()
+	app.Name = "drone-server"
+	app.Version = version.Version.String()
+	app.Usage = "drone server"
+	app.Action = server
+	app.Flags = flags
+	app.Before = before
+
+	if err := app.Run(os.Args); err != nil {
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+}
